Properties v2: extract shared PropertiesEditTab shell
The patient and invoice property tabs each reimplemented the same shell —
layout, draft, confirmed Save button, dirty reporting, and empty/loading
states — differing only in how they fetched data, resolved permissions, and
saved.

Extract that shell into a shared presentational `PropertiesEditTab` in common
(alongside `PropertyV2DetailRows`) and reduce both tabs to thin containers:

- Patient `CustomPropertiesTab` fetches its own data, gates on `PatientMutate`,
  and saves via its mutation hook.
- Invoice `InvoicePropertiesTab` resolves per-type permission + invoice status,
  sorts definitions, and wraps the save with `throwIfStructuredError`.
